infrastructure/idlharness.js/IdlInterface/test_partial_interface_of.html

<!DOCTYPE HTML>
<html>

<head>
  <meta charset="utf-8">
  <meta name="variant" content="">
  <title>idlharness: Partial interface</title>
  <script src="/infrastructure/testharness.js/resources/variant-promise.js"></script>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/WebIDLParser.js"></script>
  <script src="/resources/idlharness.js"></script>
</head>

<body>
  <p>Verify the series of sub-tests that are executed for "partial" interface objects.</p>
  <script>
    "use strict";

    // No original existence
    (() => {
      const idlArray = new IdlArray();
      idlArray.add_idls('partial interface A {};');
      idlArray.test();
    })();

    // Valid exposure (Note: Worker -> {Shared,Dedicated,Service}Worker)
    (() => {
      const idlArray = new IdlArray();
      idlArray.add_untested_idls(`
        [Exposed=(Worker)]
        interface B {};

        [Exposed=(DedicatedWorker, ServiceWorker, SharedWorker)]
        interface C {};`);
      idlArray.add_idls(`
        [Exposed=(DedicatedWorker, ServiceWorker, SharedWorker)]
        partial interface B {};

        [Exposed=(Worker)]
        partial interface C {};`);
      idlArray.collapse_partials();
    })();

    // Invalid exposure
    (() => {
      const idlArray = new IdlArray();
      idlArray.add_untested_idls(`
        [Exposed=(Window, ServiceWorker)]
        interface D {};`);
      idlArray.add_idls(`
        [Exposed=(DedicatedWorker)]
        partial interface D {};`);
      idlArray.collapse_partials();
    })();

    // Original is a namespace, not an interface.
    (() => {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
        partial interface E {};
        namespace E {};`);
      idlArray.collapse_partials();
    })();

    (() => {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
        partial interface F {};
        partial interface mixin G{};
      `);
      idlArray.add_dependency_idls(`
        interface F {};
        interface mixin G {};
        interface mixin H {};
        F includes H;
        I includes H;
        J includes G;
        interface K : J {};
        interface L : F {};
      `);
      test(() => {
        // F is tested, so H is a dep.
        assert_true('F' in idlArray.includes, 'Fs includes should be picked up');
        assert_true(idlArray.includes.F.includes('H'), 'H should be picked up');
        // H is not tested, so I is not a dep.
        assert_false('I' in idlArray.includes, 'I should be ignored');
        // G is a dep, so J is a dep.
        assert_true('J' in idlArray.includes, 'J should be picked up');
        // But, K isn't a dep just because of J.
        assert_false('K' in idlArray.members, 'K should be ignored');
        // L inherits F, so should be picked up.
        assert_false('L' in idlArray.members, 'L should be picked up');
      }, 'partial mixin dep implications');
    })();

    // Name clash (partials)
    (() => {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
        interface M { attribute any A; };
        partial interface M { attribute any A; };`);
      idlArray.collapse_partials();
    })();
  </script>
</body>

</html>
